+++ pip_services3_components/lock/NullLock.py
@@ -12,7 +12,7 @@ class NullLock(ILock):
     but shall be disabled.
     """
 
-    def try_acquire_lock(self, correlation_id: Optional[str], key: str, ttl: float) -> bool:
+    def try_acquire_lock(self, correlation_id: Optional[str], key: str, ttl: int) -> bool:
         """
         Makes a single attempt to acquire a lock by its key.
         It returns immediately a positive or negative result.
@@ -24,7 +24,7 @@ class NullLock(ILock):
         """
         return True
 
-    def acquire_lock(self, correlation_id: Optional[str], key: str, ttl: float, timeout: float):
+    def acquire_lock(self, correlation_id: Optional[str], key: str, ttl: int, timeout: int):
         """
         Releases prevously acquired lock by its key.
